Message Processing
Almost finished implementing processing of messages received from the
server, our agent holds  info about each of the messages received from
all other agents in the game in a class called AgentBelief.
This class will be queried by the future strategy to help the agent
make it's decisions.

Work left:
1. Understand how talk_number is encoded and create a db of all the
messages sent such that if AGREE 20 is sent we will know what is the
message with talk number 20.
2. Check if there is more processing to do to logical messages such
as AND NOT and OR (which is not yet implemented).
3. Check about NESTED structure of the  agents knowledge, do we need
to also hold information about messages of other agents in the belief
of an agent of a given index.