Wrap Up Processing.
Finished implementing processing of messages received from the game
server, currently all the important messages (not "Skip" or "Over") of
each agent is stored in it's AgentBelief object.
In the future the agent beliefs will be used by our strategies that will
guide the agent what to do next given the information he has soon
throughout the game.