Hostility and Fondness.
Replaced the old implementation with a new one that lets villagers
players hold an AgentPerspective class which updates the enemies and
cooperators of other agents based on sentences passing on throughout
the game.
Next features that will be implemented:
1. Perspective for werewolf - integrate processing of communication
in the night phase that is used by the werewolves.
2. Find greater signs of cooperation in sentences, current approach
doesn't work well.
3. Detect lies by matching agent perspectives.
4. Approach to inquires and requests made by other agents toward us.