Implemented recall of processed sentences.
Implemented a dictionary which contains all the sentences that
were processed, each agent perspective will hold a reference to the
object that fills the dictionary that matches talk number to processed
sentences. That way when there is an agreement or disagreement
we can recall what the sentence that the agent agreed with
"meant".

Note: Hasn't been fully tested.