Created Sentence Dissection Module
Moved the logic of dissecting sentences into a separate module
which is used by the agent perspective and can be used by any other class
that wishes to dissect sentences by itself. The result of the dissection
contains currently all information which was relevant to the agent
perspective class. If someone wishes to take more information out of
a sentence we can add a property to the dissection result and give
it a value during the dissection of the sentence.