Finished implementing game graph + visualization.
Implemented a process of building a graph of interactions in the game
in the GroupFinder class, this graph is updated based on dissection
of sentences of the players in the game.

At the end of the game the graph is drawn into a file named
"Graph.png" which is used for debugging purposes, that way we can go
over a game, see the resulting graph and  tune in new heuristics
or delete ones that don't work.

Future advances (currently not implemented):
1.Use the gathered picture that we have in the graph to find cooperations
that don't necessarily stem up  from individual sentences, like agents
requesting similar things based on the same reason.
2. External API to the graph. Currently the game graph allows you to get
any node you like and traverse over the graph. If there is any need
an implementation of BFS/DFS will be added to get connected components
of the graph.