DayOne Agent State
Wrote generator of random questions for the first day of the game
we work randomly because we don't have any information on the first day.
This state is just one of many future states that will be added in the
future which will decide what our agent will say.

A new questions might be added and old ones might be removed from
the questions pool based on their performance.