Started Implementing Task Module
Started implementing a module for tasks that will be built throughout
the game, each task has an importance that decays with time. During the
decision of what to say next the agent will choose between the tasks or
act randomly with some given probability.

The first task that was implemented was for agents that admitted to the
same roles. In this case when the task is handle we will send a sentence
that shows the lie of these players and accuse them of being on the
werewolf team.