Finished implementing the BaseState and VoterModel
Implemented the basic state of an agent, this is a state
in which we randomize between asking random questions and doing
one of our created tasks.

Also implemented the basic voter model for the villager player.
A singleton classs named PlayerEvaluation was created, it's job is to
hold a mapping between each player and how likely do we think that he
is HUMAN the closer the score is to zero the more likely he is to be
human.

Future Work:
- Add a task that requests other agents to vote to player that is very
likely to be a werewolf.
- Work on the cooperation module in agent perspective, what we have now
is not enough.